Some cleaning up, code runs quite smoothly now
