+++ src/wam2layers/utils/profiling.py
@@ -1,0 +1,19 @@
+import time
+import psutil
+
+
+class Profiler:
+    def __init__(self):
+        self.t0 = time.time()
+        self.mem0 = self._current_memory()
+
+    def __call__(self):
+        """Report memory and time increments."""
+        dt = round((time.time() - self.t0), 2)
+        dmem = self._current_memory() - self.mem0
+        # self.t0 = time.time()
+        # self.mem0 = self._current_memory()
+        return dt, dmem
+
+    def _current_memory(self):
+        return psutil.Process().memory_info().rss / (1024 * 1024)
